+++ config/.gitignore
@@ -1,2 +1,3 @@
 # Keys
 api_keys.env
+prototypes_keys.env
