+++ prototypes/backend/generation/copy_database.sh
