+++ prototypes/pyproject.toml
@@ -13,6 +13,7 @@ django-ninja = "^1.1.0"
 jinja2 = "^3.1.4"
 flask = "^3.0.3"
 openai = "^1.47.0"
+groq = "^0.15.0"
 requests = "^2.32.3"
 
 [tool.poetry.group.dev.dependencies]
